==> programs/2_HCMS.Rmd <==
---
title: "ANOVA for temperature tolerance"
author: "Emma Chandler"
date: "September 11, 2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load Libraries
```{r message=FALSE}
library(dplyr)
library(ggplot2)
library(lmerTest)
library(EnvStats)
library(ggpubr)
library(emmeans)
library(car)
```



## Load Data
```{r}
remove(list = ls())

H.CMS <- read.csv("~/Horsenettle/data/CMSdata.csv", header = TRUE)

```


## Differences between north and south for the initial measurements
```{r}

# Is there a difference between the two max damage measurements
t.test(H.CMS$exp_max, H.CMS$cntr_max)

# Mixed effects model checking difference between regions
HCMSL.origin <- lmer(exp_max ~ location + (1|ramet) + (1|ID), data = H.CMS)

summary(HCMSL.origin)
anova(HCMSL.origin)


# One-way ANOVA checking difference between genets
HCMSL.origin2 <- aov(exp_max ~ID, H.CMS)

summary.aov(HCMSL.origin2)

```



## Calculate CMS 
```{r}
# Create  proportions of damage over max
trtpro <- H.CMS$exp_trt / H.CMS$exp_max
cntrpro <- H.CMS$cntr_trt / H.CMS$cntr_max

# Calculate CMS using proportions
H.CMS$CMS <- ((1 - trtpro)/(1 - cntrpro))

# Export data frame for Correlation
# write.csv(x = H.CMS, file = "~/Horsenettle/data/processed/H.CMS.csv")

```

## Test for outliers
```{r}
# Create a boxplot of the dataset, outliers are shown as two distinct points
(outliers <- boxplot(data = H.CMS, CMS~location, plot=TRUE)$out)  

# Use number of potential outliers from previous code
rosner <- rosnerTest(H.CMS$CMS, k = 4)
rosner$all.stats
# No Outliers

```

# Creating a figure theme for all plots
```{r}
HStheme <- function(){            
  theme_bw()+                          
  theme(axis.text.x = element_text(size = 12), 
        text = element_text(family = "sans"), 
        axis.text.y = element_text(size = 12),
        plot.subtitle = element_text(size = 16, vjust = 1, hjust = 0.5),
        axis.title = element_text(size = 12),
        panel.grid = element_blank(),
        plot.margin = unit(c(0.5, 0.5, 0.5, 0.5), units = , "cm"),
        plot.title = element_text(size = 20, vjust = 1, hjust = 0.5),
        legend.text = element_text(size = 12),
        legend.title = element_blank())
}
```



# Is there variance between regions, while accounting for block, population, and genet as random effects?
```{r}
HCMS_lmer1 <- lmer(CMS ~ location + (1|ramet) + (1|ID), data = H.CMS)
AIC(HCMS_lmer1)
(HCMS_lmer1.a <- anova(HCMS_lmer1))
(HCMS_glmer1.r <- ranova(HCMS_lmer1))

HCMS_lmer1_pop <- lmer(CMS ~ location + population + (1|ramet) + (1|ID), data = H.CMS)
AIC(HCMS_lmer1_pop)
(HCMS_lmer1_pop.a <- anova(HCMS_lmer1_pop))
(HCMS_lmer1_pop.r <- ranova(HCMS_lmer1_pop))


summary(HCMS_lmer1)

# Include populations as a random effect with southern populations pooled
H.CMS <- H.CMS %>% mutate(population_2 = case_when(location == "South" ~ "South",
                                                   population == "Prairie Island" ~ "Prairie Island",
                                                   population == "Frontenac" ~ "Frontenac"))

HCMS_lmer1_pop <- lmer(CMS ~ location + (1|ramet) + (1|population_2/ID), data = H.CMS)
AIC(HCMS_lmer1_pop)

(HCMS_lmer1_pop.a <- anova(HCMS_lmer1_pop))
(HCMS_lmer1_pop.r <- ranova(HCMS_lmer1_pop))

summary(HCMS_lmer1_pop)

# Data Visualization
(H.CMS.p1 <- ggplot(H.CMS, aes(ramet, CMS, fill = location)) + 
    geom_boxplot(alpha = 0.8) +
    scale_fill_manual(values = c("#104E8B", "#CD2626")) +
    HStheme() +
    labs(x = "Block", y = "CMS Ratio", title = "Cell Membrane Stability", subtitle = "Hot Treatment"))

# Saving graph
# ggsave("figures/H.CMS.p1.png", width = 7, height = 5, dpi = 300)

(H.CMS.p1 <- ggplot(H.CMS, aes(location, CMS, fill = location)) + 
    geom_boxplot(alpha = 0.8) +
    scale_fill_manual(values = c("#104E8B", "#CD2626")) +
    HStheme() +
    labs(x = "Block", y = "CMS Ratio", title = "Cell Membrane Stability", subtitle = "Hot Treatment"))


# Saving graph
# ggsave("figures/H.CMS.p1b.png", width = 7, height = 5, dpi = 300)
```



## Is there a difference in CMS between the blocks
```{r}
HCMS_block <- aov(CMS ~ ramet, H.CMS)
summary(HCMS_block)


posthoc <- TukeyHSD(x=HCMS_block, 'ramet', conf.level=0.95)
posthoc


(CMS.block.plot <- ggplot(H.CMS, aes(x = ramet, y = CMS, fill = location)) + 
    geom_boxplot(alpha = 0.8) +
    scale_fill_manual(values = c("#104E8B", "#CD2626")) +
    HStheme() +
    labs(x = "Block", y = "CMS Ratio"))

# Saving graph
# ggsave("figures/H.CMS.BLOCK.png", width = 7, height = 5, dpi = 300)

```
Since there is variation between blocks that is probably due to temperatures in the greenhouse affecting the the plants, is there a significant difference between north and south for individual ramets
```{r}

# Subset data from south
south <- subset(H.CMS,  location == "South",
                 drop = TRUE)
# subset data from north
north <- subset(H.CMS,  location == "North",
                 drop = TRUE)

# Ramet A
ramA_south <- filter(south, ramet == "A") 
ramA_north <- filter(north, ramet == "A")

ramA.ttest <- t.test(ramA_south$CMS, ramA_north$CMS, paired = TRUE)
ramA.ttest

(rametA <- H.CMS %>% 
    filter(ramet == "A") %>% 
    ggplot(aes(location, CMS, fill = location)) + 
    geom_boxplot(alpha = 0.8) +
    scale_fill_manual(values = c("#104E8B", "#CD2626")) +
    HStheme() +
    labs(x = "Region", y = "CMS Ratio", title = "Ramet A"))

# Saving graph
# ggsave("figures/H.CMS_rametA.png", width = 7, height = 5, dpi = 300)


# Ramet B
ramB_south <- filter(south, ramet == "B") 
ramB_north <- filter(north, ramet == "B")

ramB.ttest <- t.test(ramB_south$CMS, ramB_north$CMS, paired = TRUE)
ramB.ttest


# Ramet C
ramC_south <- filter(south, ramet == "C") 
ramC_north <- filter(north, ramet == "C")

ramC.ttest <- t.test(ramC_south$CMS, ramC_north$CMS, paired = TRUE)
ramC.ttest


# Ramet D
ramD_south <- filter(south, ramet == "D") 
ramD_north <- filter(north, ramet == "D")

ramD.ttest <- t.test(ramD_south$CMS, ramD_north$CMS, paired = TRUE)
ramD.ttest
```


## Is there variation in temperature tolerance between genets with ramet as replicates
```{r}
HCMS_aov2 <- aov(CMS ~ ID, data = H.CMS)

summary(HCMS_aov2)
(HCMS_lmer2.a <- anova(HCMS_aov2))

TUKEY <- TukeyHSD(HCMS_aov2)
plot(TUKEY, las = 1)

# #Crate summarised data
# detach(package:plyr)

summary <- H.CMS %>% 
  group_by(ID, location) %>% 
  summarise(mean = mean(CMS),se = sd(CMS)/(sqrt(length(CMS))))

# dot plot
(H.CMS.dot <- ggplot(summary, aes(x = reorder(ID, mean), y = mean)) +
    geom_errorbar(aes(ymin = mean - se, ymax = mean + se), width = 0.2) +
    geom_point(size = 3, aes(colour = location)) +
    scale_color_manual(values = c("#104E8B", "#CD2626"))+
    HStheme()+
    stat_compare_means(label = "p.signif", method = "t.test",ref.group = ".all.") + 
    theme(axis.text.x = element_text(size = 10, angle = 90, vjust = 0.5)) +
    labs(x = "Genotype", y = "CMS Ratio", title = "Cell Membrane Stability", subtitle = "Hot Treatment"))


(H.CMS.dot <- ggplot(H.CMS, aes(x = reorder(ID, CMS, FUN = median, na.rm = TRUE), y = CMS, fill = location)) +
    geom_boxplot(alpha = 0.8) +
    HStheme() +
    scale_fill_manual(values = c("#104E8B", "#B22222"))+
    theme(axis.text.x = element_text(angle = 90, size = 8, vjust = 0.5),
          axis.text.y = element_text(size = 8))+
    labs(x = "Genet", y = "CMS Ratio", title = "Cell Membrane Stability", subtitle = "Hot Treatment"))


# Saving graph
ggsave("figures/H.CMS.p2.png", width = 7, height = 5, dpi = 300)

```


## Is there variation in genets for the individual regions?
```{r}
# Is there variation in the individuals with in the independent Regions

#North
H.CMS_N <- H.CMS %>% 
  filter(location == "North")

HCMS_N <- aov(CMS ~ ID, data = H.CMS_N)
summary(HCMS_N)


#South
H.CMS_S <- H.CMS %>% 
  filter(location == "South")

HCMS_S <- aov(CMS ~ ID, data = H.CMS_S)
summary(HCMS_S)



```




# Variance between groups

```{r}
reg <- bartlett.test(CMS ~ location, data = H.CMS)
reg

```


# 1 ) Is there variance between populations?
```{r}

# Mixed Model for populations
pop.aov <- lmer(CMS ~ population + (1|ramet), data = H.CMS)

(pop.aov.a <- anova(pop.aov))
(pop.aov.r <- ranova(pop.aov))


# Posthoc Test
PH.pop <-emmeans(pop.aov, list(pairwise ~ population), adjust = "tukey")
plot(PH.pop)


my_comparisons <- list(c("Cemetary", "Oil Patch"), c("Cemetary", "Prairie Island"), c("Cemetary", "Frontenac"),   c("Cemetary", "Reserve"), c("Oil Patch", "Prairie Island"), c("Frontenac", "Oil Patch"), c("Oil Patch", "Reserve"), c("Frontenac", "Prairie Island"),   c("Prairie Island", "Reserve"), c("Frontenac", "Reserve")) # comparisons for post-hoc tests
# Edit until here


ggboxplot(H.CMS, x = "population", y = "CMS",
          fill = "location") +
  stat_compare_means(comparisons = my_comparisons, label = "p.signif") +
  scale_fill_manual(values = c("#104E8B", "#CD2626")) +
  HStheme() +
  theme(axis.text.x = element_text(size = 12, angle = 0)) +
  labs(x = "Population", y = "CMS")



# Saving graph
# ggsave("figures/H.CMS.POPULATION_sig.png", width = 7, height = 5, dpi = 300)

```
