==> programs/11_Pollen_TPC.Rmd <==
---
title: "Pollen Tube Analysis"
author: "Emma Chandler"
date: "October 27, 2021"
output: html_document
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
```

## Load Libraries

```{r message=FALSE} 

library(purrr)
library(rTPC)
library(nls.multstart)
library(tidyr)
library(dplyr)
library(broom)
library(ggplot2)
library(ggpubr)
```

## Load Data

```{r}
remove(list = ls())

Pollen <- read.csv("~/Horsenettle/data/pollenTube_NEW.csv", header = TRUE)

Pollen$fullID <- paste(Pollen$ID, Pollen$ramet)

```




## Calculations

```{r}

# Calculating germination percentage
Pollen$germ <- Pollen$tubes/Pollen$total_grains * 100


# Calculating pollen tube growth rate
Pollen$PTGR <- Pollen$mean_length/Pollen$time

```


## Creating a figure theme for all plots
```{r}
HStheme <- function(){            
  theme_bw()+                          
  theme(axis.text.x = element_text(size = 12), 
        text = element_text(family = "sans"), 
        axis.text.y = element_text(size = 12),
        plot.subtitle = element_text(size = 16, vjust = 1, hjust = 0.5),
        axis.title = element_text(size = 12),
        panel.grid = element_blank(),
        plot.margin = unit(c(0.5, 0.5, 0.5, 0.5), units = , "cm"),
        plot.title = element_text(size = 14, vjust = 1, hjust = 0.5),
        legend.text = element_text(size = 12),
        legend.title = element_blank())
}
```



## Pollen Germination
Fitting curves using rTPC fit for thermal performance curves using nls.multstart
Author Daniel Padfield (12-3-2021)
https://padpadpadpad.github.io/rTPC/articles/rTPC.html 

```{r, message=FALSE}

get_model_names()

# Create nested data frames for each plant
by_ID <- Pollen %>% 
  group_by(fullID) %>% 
  nest()

# Save a function with the curve fitting model
curves <- function(df) {
  nls_multstart(germ~quadratic_2008(temp = temp, a,b,c),
                     data = df,
                     iter = 500, 
                     start_lower = get_start_vals(df$temp, df$germ,
                                                     model_name = 'quadratic_2008') - 10,
                        start_upper = get_start_vals(df$temp, df$germ,
                                                     model_name = 'quadratic_2008') + 10,
                        lower = get_lower_lims(df$temp, df$germ, 
                                               model_name = 'quadratic_2008'),
                        upper = get_upper_lims(df$temp, df$germ, 
                                               model_name = 'quadratic_2008'),
                     supp_errors = "Y")
  }


# Run the model for each plant and save to the by_ID data frame
by_ID <- by_ID %>% 
  mutate(model = map(data,curves))


params <- by_ID %>% 
  select(model) %>% 
  calc_params()


# Function to calculate the parameters of the curves
param <- function(df) {
  calc_params(model)
}

# Calculate parameters for all curves
by_ID <- by_ID %>% 
  mutate(param = map(model, calc_params))

# Unnest parameters into one dataframe
params <- unnest(by_ID, param) %>% 
  select(-data, -model)


# Add regional information to the parameters data frame
region_dat <- Pollen %>% 
  select(1:5, 16) %>% 
  distinct()

params_final <- merge(params, region_dat, by = "fullID")


# Create csv with the parameters
# write.csv(params_final, "~/Horsenettle/Data/processed/germParams.csv")


```


## Mean Pollen Tube Growth Rate

```{r, message=FALSE}

get_model_names()

# Create nested data frames for each plant
by_ID.pt <- Pollen %>% 
  group_by(fullID) %>% 
  nest()

# Save a function with the curve fitting model
curves.pt <- function(df) {
  nls_multstart(PTGR~quadratic_2008(temp = temp, a,b,c),
                     data = df,
                     iter = 500, 
                     start_lower = get_start_vals(df$temp, df$PTGR,
                                                     model_name = 'quadratic_2008') - 10,
                        start_upper = get_start_vals(df$temp, df$PTGR,
                                                     model_name = 'quadratic_2008') + 10,
                        lower = get_lower_lims(df$temp, df$PTGR, 
                                               model_name = 'quadratic_2008'),
                        upper = get_upper_lims(df$temp, df$PTGR, 
                                               model_name = 'quadratic_2008'),
                     supp_errors = "Y")
  }


# Run the model for each plant and save to the by_ID data frame
by_ID.pt <- by_ID.pt %>% 
  mutate(model = map(data,curves.pt))


# Calculate parameters for all curves
by_ID.pt <- by_ID.pt %>% 
  mutate(param = map(model, calc_params))

# Unnest parameters into one dataframe
params.pt <- unnest(by_ID.pt, param) %>% 
  select(-data, -model)


# Add regional information to the parameters data frame
params_final.pt <- merge(params.pt, region_dat, by = "fullID")


# Create csv with the parameters
# write.csv(params_final.pt, "~/Horsenettle/Data/processed/ptgrParams.csv")


```




## Single plant with plot - Germination

```{r}

# Plotting one curve
single <- filter(Pollen, fullID == "PI1 A")

# get starting values
start_vals <- get_start_vals(single$temp, single$germ, model_name = 'quadratic_2008')
low_lims <- get_lower_lims(single$temp, single$germ, model_name = 'quadratic_2008')
upper_lims <- get_upper_lims(single$temp, single$germ, model_name = 'quadratic_2008')

start_vals
low_lims

fit <- nls_multstart(germ~quadratic_2008(temp = temp, a,b,c),
                     data = single,
                     iter = 500, 
                     start_lower = start_vals - 10,
                     start_upper = start_vals + 10,
                     lower = low_lims,
                     upper = upper_lims,
                     supp_errors = "Y")

parameters <- calc_params(fit) %>% 
  mutate_all(round, 2)

parameters$ID <- "OP1 A"
parameters$location <- "North"



# Predict new data
new_data <- data.frame(temp = seq(min(single$temp),
                                  max(single$temp),
                                  0.5))




preds_PI <- augment(fit, newdata = new_data)

# Plot data and model fit
(PI1A.p <- ggplot(single, aes(temp,germ)) +
  geom_point() +
  geom_line(aes(temp, .fitted), preds_PI, col = 'blue') +
  HStheme() +
  labs(x = 'Temperature (ºC)',
       y = '% Germination',
       title = "PI1 A"))

ggsave("figures/germ_curve_N.png", width = 7, height = 5, dpi = 300)




```

## Single plant with plot

```{r}

# Plotting one curve
single <- filter(Pollen, fullID == "OP1 A")

# get starting values
start_vals <- get_start_vals(single$temp, single$germ, model_name = 'quadratic_2008')
low_lims <- get_lower_lims(single$temp, single$germ, model_name = 'quadratic_2008')
upper_lims <- get_upper_lims(single$temp, single$germ, model_name = 'quadratic_2008')

start_vals
low_lims

fit <- nls_multstart(germ~quadratic_2008(temp = temp, a,b,c),
                     data = single,
                     iter = 500, 
                     start_lower = start_vals - 10,
                     start_upper = start_vals + 10,
                     lower = low_lims,
                     upper = upper_lims,
                     supp_errors = "Y")

parameters <- calc_params(fit) %>% 
  mutate_all(round, 2)

parameters$ID <- "OP1 A"
parameters$location <- "South"



# Predict new data
new_data <- data.frame(temp = seq(min(single$temp),
                                  max(single$temp),
                                  0.5))




preds_OP <- augment(fit, newdata = new_data)

# Plot data and model fit
(OP1A.p <- ggplot(single, aes(temp,germ)) +
  geom_point() +
  geom_line(aes(temp, .fitted), preds_OP, col = 'red') +
  HStheme() +
  labs(x = "Temperature (ºC)", y = '% Germination',
       title = "OP1 A"))

ggsave("figures/germ_curve_S.png", width = 7, height = 5, dpi = 300)

(stacked_quad <- ggpubr::ggarrange(OP1A.p, PI1A.p, ncol = 1, nrow = 2))


# ggsave("figures/stacked_genet.png", width = 4, height = 5, dpi = 300)

```



## Single plant with plot - PTGR

```{r}

# Plotting one curve
single <- filter(Pollen, fullID == "PI1 A")

# get starting values
start_vals <- get_start_vals(single$temp, single$PTGR, model_name = 'quadratic_2008')
low_lims <- get_lower_lims(single$temp, single$PTGR, model_name = 'quadratic_2008')
upper_lims <- get_upper_lims(single$temp, single$PTGR, model_name = 'quadratic_2008')

start_vals
low_lims

fit <- nls_multstart(PTGR~quadratic_2008(temp = temp, a,b,c),
                     data = single,
                     iter = 500, 
                     start_lower = start_vals - 10,
                     start_upper = start_vals + 10,
                     lower = low_lims,
                     upper = upper_lims,
                     supp_errors = "Y")

parameters <- calc_params(fit) %>% 
  mutate_all(round, 2)

parameters$ID <- "OP1 A"
parameters$location <- "North"



# Predict new data
new_data <- data.frame(temp = seq(min(single$temp),
                                  max(single$temp),
                                  0.5))




preds_PI <- augment(fit, newdata = new_data)

# Plot data and model fit
(PI1A.p <- ggplot(single, aes(temp,PTGR)) +
  geom_point() +
  geom_line(aes(temp, .fitted), preds_PI, col = 'blue') +
  HStheme() +
  labs(x = 'Temperature (ºC)',
       y = 'PTGR',
       title = "PI1 A"))

ggsave("figures/PTGR_curve_N.png", width = 7, height = 5, dpi = 300)




```

## Single plant with plot

```{r}

# Plotting one curve
single <- filter(Pollen, fullID == "OP1 A")

# get starting values
start_vals <- get_start_vals(single$temp, single$PTGR, model_name = 'quadratic_2008')
low_lims <- get_lower_lims(single$temp, single$PTGR, model_name = 'quadratic_2008')
upper_lims <- get_upper_lims(single$temp, single$PTGR, model_name = 'quadratic_2008')

start_vals
low_lims

fit <- nls_multstart(PTGR~quadratic_2008(temp = temp, a,b,c),
                     data = single,
                     iter = 500, 
                     start_lower = start_vals - 10,
                     start_upper = start_vals + 10,
                     lower = low_lims,
                     upper = upper_lims,
                     supp_errors = "Y")

parameters <- calc_params(fit) %>% 
  mutate_all(round, 2)

parameters$ID <- "OP1 A"
parameters$location <- "South"



# Predict new data
new_data <- data.frame(temp = seq(min(single$temp),
                                  max(single$temp),
                                  0.5))




preds_OP <- augment(fit, newdata = new_data)

# Plot data and model fit
(OP1A.p <- ggplot(single, aes(temp,PTGR)) +
  geom_point() +
  geom_line(aes(temp, .fitted), preds_OP, col = 'red') +
  HStheme() +
  labs(x = "Temperature (ºC)", y = 'PTGR',
       title = "OP1 A"))

ggsave("figures/PTGR_curve_S.png", width = 7, height = 5, dpi = 300)

(stacked_quad <- ggpubr::ggarrange(OP1A.p, PI1A.p, ncol = 1, nrow = 2))


# ggsave("figures/stacked_genet.png", width = 4, height = 5, dpi = 300)
```
